FoxTruyen: Mirate to 1.6
