+++ src/vi/truyengg/build.gradle.kts
@@ -8,7 +8,7 @@ keiyoushi {
     name = "FoxTruyen"
     versionCode = 13
     contentWarning = ContentWarning.SAFE
-    libVersion = "1.4"
+    libVersion = "1.6"
 
     source {
         lang = "vi"
